moved functions from Objhandler to inihandler
